boss

+++ controllers/educationController.js
@@ -9,38 +9,78 @@ import cloudinary from "cloudinary";
 import { User } from "../models/User.js";
 
 export const createEducationGraduation = catchAsyncError(async (req, res, next) => {
-    const { university, specialization, yearOfPassing, percentageOfMarks, classGrade } = req.body;
-    const file = req.file;
-
-    const fileUri = getDataUri(file);
-
-    const mycloud = await cloudinary.v2.uploader.upload(fileUri.content);
+    let { university, specialization, yearOfPassing, percentageOfMarks, classGrade } = req.body;
+    let file = null;
 
     const user = await User.findById(req.user._id).populate("userData.education.graduation");
 
-    if (user.userData.education.graduation) {
-        await cloudinary.v2.uploader.destroy(user.userData.education.graduation.selfAttestedCopy.public_id);
-        await Graduation.deleteOne({ _id: user.userData.education.graduation._id });
+    if (req.file) {
+        console.log('upper');
+        file = req.file;
+
+
+        const fileUri = getDataUri(file);
+
+        const mycloud = await cloudinary.v2.uploader.upload(fileUri.content);
+
+
+        if (user.userData.education.graduation) {
+            if (user.userData.education.graduation.selfAttestedCopy.public_id) {
+                await cloudinary.v2.uploader.destroy(user.userData.education.graduation.selfAttestedCopy.public_id);
+                await Graduation.deleteOne({ _id: user.userData.education.graduation._id });
+            }
+        }
+
+        user.userData.education.graduation = await Graduation.create({
+            university,
+            specialization,
+            yearOfPassing,
+            percentageOfMarks,
+            classGrade,
+            selfAttestedCopy: {
+                public_id: mycloud.public_id,
+                url: mycloud.secure_url,
+            },
+        })
+
+        await user.save();
+        res.status(201).json({
+            success: true,
+            user,
+            message: "Graduation Details Saved Successfully",
+        });
+    }
+    else {
+        console.log('lower')
+
+        if (user.userData.education.graduation) {
+            if (user.userData.education.graduation.selfAttestedCopy.public_id) {
+                await cloudinary.v2.uploader.destroy(user.userData.education.graduation.selfAttestedCopy.public_id);
+                await Graduation.deleteOne({ _id: user.userData.education.graduation._id });
+            }
+        }
+
+        user.userData.education.graduation = await Graduation.create({
+            university,
+            specialization,
+            yearOfPassing,
+            percentageOfMarks,
+            classGrade,
+            selfAttestedCopy: {
+                public_id: "Not Provided",
+                url: "Not Provided",
+            },
+        })
+        await user.save();
+        res.status(201).json({
+            success: true,
+            user,
+            message: "Redirected to Post Graduation Details",
+        });
     }
 
-    user.userData.education.graduation = await Graduation.create({
-        university,
-        specialization,
-        yearOfPassing,
-        percentageOfMarks,
-        classGrade,
-        selfAttestedCopy: {
-            public_id: mycloud.public_id,
-            url: mycloud.secure_url,
-        },
-    })
 
-    await user.save();
-    res.status(201).json({
-        success: true,
-        user,
-        message: "Graduation Details Saved Successfully",
-    });
+
 });
 
 export const getGraduationDetails = catchAsyncError(async (req, res, next) => {
@@ -59,38 +99,72 @@ export const getGraduationDetails = catchAsyncError(async (req, res, next) => {
 });
 
 export const createEducationPostGraduation = catchAsyncError(async (req, res, next) => {
-    const { university, specialization, yearOfPassing, percentageOfMarks, classGrade } = req.body;
-    const file = req.file;
-
-    const fileUri = getDataUri(file);
-
-    const mycloud = await cloudinary.v2.uploader.upload(fileUri.content);
+    let { university, specialization, yearOfPassing, percentageOfMarks, classGrade } = req.body;
+    let file = null;
 
     const user = await User.findById(req.user._id).populate("userData.education.postGraduation");
 
     if (user.userData.education.postGraduation) {
-        await cloudinary.v2.uploader.destroy(user.userData.education.postGraduation.selfAttestedCopy.public_id);
-        await PostGraduation.deleteOne({ _id: user.userData.education.postGraduation._id });
+        if (user.userData.education.postGraduation.selfAttestedCopy.public_id) {
+            await cloudinary.v2.uploader.destroy(user.userData.education.postGraduation.selfAttestedCopy.public_id);
+            await PostGraduation.deleteOne({ _id: user.userData.education.postGraduation._id });
+        }
     }
 
-    user.userData.education.postGraduation = await PostGraduation.create({
-        university,
-        specialization,
-        yearOfPassing,
-        percentageOfMarks,
-        classGrade,
-        selfAttestedCopy: {
-            public_id: mycloud.public_id,
-            url: mycloud.secure_url,
-        },
-    })
+    if (req.file) {
+        console.log('upper');
+        file = req.file;
 
-    await user.save();
-    res.status(201).json({
-        success: true,
-        user,
-        message: "PostGraduation Details Saved Successfully",
-    });
+
+        const fileUri = getDataUri(file);
+
+        const mycloud = await cloudinary.v2.uploader.upload(fileUri.content);
+
+
+
+
+
+        user.userData.education.postGraduation = await PostGraduation.create({
+            university,
+            specialization,
+            yearOfPassing,
+            percentageOfMarks,
+            classGrade,
+            selfAttestedCopy: {
+                public_id: mycloud.public_id,
+                url: mycloud.secure_url,
+            },
+        })
+
+        await user.save();
+        res.status(201).json({
+            success: true,
+            user,
+            message: "PostGraduation Details Saved Successfully",
+        });
+    }
+    else {
+        console.log('lower')
+
+        user.userData.education.postGraduation = await PostGraduation.create({
+            university,
+            specialization,
+            yearOfPassing,
+            percentageOfMarks,
+            classGrade,
+            selfAttestedCopy: {
+                public_id: "Not Provided",
+                url: "Not Provided",
+            },
+        })
+        await user.save();
+        res.status(201).json({
+            success: true,
+            user,
+            message: "Redirected to MPhil Details",
+        });
+    }
+
 });
 
 export const getPostGraduationDetails = catchAsyncError(async (req, res, next) => {
@@ -111,37 +185,65 @@ export const getPostGraduationDetails = catchAsyncError(async (req, res, next) =
 
 export const createEducationMphil = catchAsyncError(async (req, res, next) => {
     const { university, specialization, yearOfPassing, percentageOfMarks, classGrade } = req.body;
-    const file = req.file;
-
-    const fileUri = getDataUri(file);
-
-    const mycloud = await cloudinary.v2.uploader.upload(fileUri.content);
-
     const user = await User.findById(req.user._id).populate("userData.education.mPhil");
 
     if (user.userData.education.mPhil) {
-        await cloudinary.v2.uploader.destroy(user.userData.education.mPhil.selfAttestedCopy.public_id);
-        await MPhil.deleteOne({ _id: user.userData.education.mPhil._id });
+        if (user.userData.education.mPhil.selfAttestedCopy.public_id) {
+            await cloudinary.v2.uploader.destroy(user.userData.education.mPhil.selfAttestedCopy.public_id);
+            await MPhil.deleteOne({ _id: user.userData.education.mPhil._id });
+        }
     }
 
-    user.userData.education.mPhil = await MPhil.create({
-        university,
-        specialization,
-        yearOfPassing,
-        percentageOfMarks,
-        classGrade,
-        selfAttestedCopy: {
-            public_id: mycloud.public_id,
-            url: mycloud.secure_url,
-        },
-    })
+    if (req.file) {
+        const file = req.file;
 
-    await user.save();
-    res.status(201).json({
-        success: true,
-        user,
-        message: "Mphil Details Saved Successfully",
-    });
+        const fileUri = getDataUri(file);
+
+        const mycloud = await cloudinary.v2.uploader.upload(fileUri.content);
+
+
+
+
+        user.userData.education.mPhil = await MPhil.create({
+            university,
+            specialization,
+            yearOfPassing,
+            percentageOfMarks,
+            classGrade,
+            selfAttestedCopy: {
+                public_id: mycloud.public_id,
+                url: mycloud.secure_url,
+            },
+        })
+
+        await user.save();
+        res.status(201).json({
+            success: true,
+            user,
+            message: "Mphil Details Saved Successfully",
+        });
+    }
+    else {
+        user.userData.education.mPhil = await MPhil.create({
+            university,
+            specialization,
+            yearOfPassing,
+            percentageOfMarks,
+            classGrade,
+            selfAttestedCopy: {
+                public_id: "Not Provided",
+                url: "Not Provided",
+            },
+        })
+
+        await user.save();
+        res.status(201).json({
+            success: true,
+            user,
+            message: "Redirected to Phd Details",
+        });
+    }
+
 });
 
 export const getMphilDetails = catchAsyncError(async (req, res, next) => {
@@ -159,37 +261,61 @@ export const getMphilDetails = catchAsyncError(async (req, res, next) => {
 
 export const createEducationPhd = catchAsyncError(async (req, res, next) => {
     const { university, specialization, yearOfPassing, percentageOfMarks, classGrade } = req.body;
-    const file = req.file;
-
-    const fileUri = getDataUri(file);
-
-    const mycloud = await cloudinary.v2.uploader.upload(fileUri.content);
-
     const user = await User.findById(req.user._id).populate("userData.education.phd");
 
     if (user.userData.education.phd) {
-        await cloudinary.v2.uploader.destroy(user.userData.education.phd.selfAttestedCopy.public_id);
-        await Phd.deleteOne({ _id: user.userData.education.phd._id });
+        if (user.userData.education.phd.selfAttestedCopy.public_id) {
+            await cloudinary.v2.uploader.destroy(user.userData.education.phd.selfAttestedCopy.public_id);
+            await Phd.deleteOne({ _id: user.userData.education.phd._id });
+        }
     }
+    if (req.file) {
+        const file = req.file;
 
-    user.userData.education.phd = await Phd.create({
-        university,
-        specialization,
-        yearOfPassing,
-        percentageOfMarks,
-        classGrade,
-        selfAttestedCopy: {
-            public_id: mycloud.public_id,
-            url: mycloud.secure_url,
-        },
-    })
+        const fileUri = getDataUri(file);
 
-    await user.save();
-    res.status(201).json({
-        success: true,
-        user,
-        message: "Phd Details Saved Successfully",
-    });
+        const mycloud = await cloudinary.v2.uploader.upload(fileUri.content);
+
+
+        user.userData.education.phd = await Phd.create({
+            university,
+            specialization,
+            yearOfPassing,
+            percentageOfMarks,
+            classGrade,
+            selfAttestedCopy: {
+                public_id: mycloud.public_id,
+                url: mycloud.secure_url,
+            },
+        })
+
+        await user.save();
+        res.status(201).json({
+            success: true,
+            user,
+            message: "Phd Details Saved Successfully",
+        });
+    }
+    else {
+        user.userData.education.phd = await Phd.create({
+            university,
+            specialization,
+            yearOfPassing,
+            percentageOfMarks,
+            classGrade,
+            selfAttestedCopy: {
+                public_id: "Not Provided",
+                url: "Not Provided",
+            },
+        })
+
+        await user.save();
+        res.status(201).json({
+            success: true,
+            user,
+            message: "Redirected to Phd Details",
+        });
+    }
 });
 
 export const getPhdDetails = catchAsyncError(async (req, res, next) => {
